+++ R/validate_templates.R
@@ -501,6 +501,18 @@ validate_templates <- function(fun.name, x){
             call. = F)
         }
       }
+      
+      # publisher - Only one publisher is allowed and the first will be used.
+      use_i <- tolower(x$template$personnel.txt$content$role) == "publisher"
+      if (sum(use_i) > 1) {
+        warning(
+          "personnel.txt has more than one 'publisher'. Only the first will ",
+          "be used.", call. = F)
+        use_i <- min(which(x$template$personnel.txt$content$role == "publisher"))
+        use_i <- which(x$template$personnel.txt$content$role == "publisher")[
+          which(x$template$personnel.txt$content$role == "publisher") != use_i]
+        x$template$personnel.txt$content <- x$template$personnel.txt$content[-c(use_i), ]
+      }
 
     }